consensus finder is working
It turns out the formulae from the paper indeed are wrong.
They couldn't be bothered checking that the algorithm is correct,
but did make sure every gender-neutral singular pronoun appeared as
"she".

Did away with the idea of writing a client benchmarking program in
C/C++. The reason was to avoid having to write more Java, but it won't
look good on the report.

Hence, I added command line options using Apache CLI. Started
implementing the benchmarker class.